refactor: 라우터 history api 적용
