+++ client/src/Components/organism/historyList/index.ts
@@ -1,6 +1,6 @@
 import Component, { PropsType, StateType } from '@/core/Component';
 import jsx from '@/core/jsx';
-import { getState, subscribe } from '@/utils/observer';
+import { getState, subscribe } from '@/core/observer';
 import { userState } from '@/Model';
 import DayList from '@/Components/atom/DayList';
 import List, { ListProps } from '@/Components/molecule/list';
