+++ client/src/utils/helpers.ts
@@ -1,0 +1,7 @@
+export const customEventEmitter = (eventType: string, detail?: object) => {
+  document.dispatchEvent(
+    new CustomEvent(eventType, {
+      detail,
+    })
+  );
+};
